wip: explicit reference tracking
This implements a potential way of identifying the impact of export
shenanigans, and possibly also file move operations (we haven't
considered those before and I haven't given it enough thought to
determine if this was even a problem before).

The implementation here is up for discussion; it's mostly a quick first
attempt to get a feeling of how it could work.